yatube/requirements.txt
certifi==2019.9.11        # via requests
chardet==3.0.4            # via requests
django==2.2.6
django-environ==0.8.1
idna==2.8                 # via requests
pillow
pyparsing==2.4.6          # via packaging
pytz==2019.3              # via django
requests==2.22.0
six==1.14.0               # via packaging
sorl-thumbnail==12.6.3
sqlparse==0.3.0           # via django
urllib3==1.25.6           # via requests
zipp==2.2.0               # via importlib-metadata
mixer==7.1.2
django-debug-toolbar==3.2.1
sentry-sdk==1.4.3
djangorestframework==3.12.4
psycopg2-binary==2.8.6
gunicorn==20.0.4
